to float

+++ src/estimator.py
@@ -2,20 +2,20 @@ def estimator(data):
   impact = {}  # your best case estimation
   severe_impact = {}  # your severe case estimation
   # import pdb; pdb.set_trace()
-  total_hospital_beds = int(0.35 * data['totalHospitalBeds'])
-  time_to_elapse = int(2 ** (int(data['timeToElapse'] / 3)))
+  total_hospital_beds = (0.35 * data['totalHospitalBeds'])
+  time_to_elapse = (2 ** (data['timeToElapse'] / 3))
 
   # currentlyInfected
-  impact['currentlyInfected'] = int(data['reportedCases'] * 10)
-  severe_impact['currentlyInfected'] = int(data['reportedCases'] * 50)
+  impact['currentlyInfected'] = (data['reportedCases'] * 10)
+  severe_impact['currentlyInfected'] = (data['reportedCases'] * 50)
 
   # infectionsByRequestedTime
-  impact['infectionsByRequestedTime'] = int(impact['currentlyInfected'] * time_to_elapse)
-  severe_impact['infectionsByRequestedTime'] = int(severe_impact['currentlyInfected'] * time_to_elapse)
+  impact['infectionsByRequestedTime'] = (impact['currentlyInfected'] * time_to_elapse)
+  severe_impact['infectionsByRequestedTime'] = (severe_impact['currentlyInfected'] * time_to_elapse)
 
   # # severeCasesByRequestedTime
-  # impact['severeCasesByRequestedTime'] = int(0.15 * impact['infectionsByRequestedTime'])
-  # severe_impact['severeCasesByRequestedTime'] = int(0.15 * severe_impact[
+  # impact['severeCasesByRequestedTime'] = (0.15 * impact['infectionsByRequestedTime'])
+  # severe_impact['severeCasesByRequestedTime'] = (0.15 * severe_impact[
   #                                                 'infectionsByRequestedTime'])
   #
   # # total_hospital_beds
@@ -25,18 +25,18 @@ def estimator(data):
   #                                                 'severeCasesByRequestedTime']
   #
   # # casesForICUByRequestedTime
-  # impact['casesForICUByRequestedTime'] = int(0.05 * impact['infectionsByRequestedTime'])
-  # severe_impact['casesForICUByRequestedTime'] = int(0.05 * severe_impact[
+  # impact['casesForICUByRequestedTime'] = (0.05 * impact['infectionsByRequestedTime'])
+  # severe_impact['casesForICUByRequestedTime'] = (0.05 * severe_impact[
   #                                                'infectionsByRequestedTime'])
   #
   # # casesForVentilatorsByRequestedTime
-  # impact['casesForVentilatorsByRequestedTime'] = int(0.02 * impact['infectionsByRequestedTime'])
-  # severe_impact['casesForVentilatorsByRequestedTime'] = int(0.02 * severe_impact[
+  # impact['casesForVentilatorsByRequestedTime'] = (0.02 * impact['infectionsByRequestedTime'])
+  # severe_impact['casesForVentilatorsByRequestedTime'] = (0.02 * severe_impact[
   #                                                        'infectionsByRequestedTime'])
   #
   # # dollarsInFlight
-  # impact['dollarsInFlight'] = int((impact['infectionsByRequestedTime'] * 0.65 * 1.5) / 30)
-  # severe_impact['dollarsInFlight'] = int((severe_impact[
+  # impact['dollarsInFlight'] = ((impact['infectionsByRequestedTime'] * 0.65 * 1.5) / 30)
+  # severe_impact['dollarsInFlight'] = ((severe_impact[
   #                                        'infectionsByRequestedTime'] * 0.65 * 1.5) / 30)
 
   output = {
